requirements.txt
# RentAByte – Python Backend Dependencies

# ── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9     # required for file upload (UploadFile)

# ── Data Validation ──────────────────────────────────────────────────────────
pydantic==2.7.1

# ── Blockchain ───────────────────────────────────────────────────────────────
web3==6.18.0                 # Polygon / Ethereum RPC interaction

# ── Encryption ───────────────────────────────────────────────────────────────
cryptography==42.0.7         # Fernet symmetric encryption

# ── Dropbox SDK ──────────────────────────────────────────────────────────────
dropbox==12.0.2

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1         # load .env file for local dev

# ── Database ────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9